Refactor: 관련 엔티티끼리 하나의 디렉토리로 묶기
- Member, Board로 나누어 묶어서 관리

Resolve: #15
