+++ src/main/java/wegrus/clubwebsite/dto/member/MemberAndJwtDto.java
@@ -3,7 +3,7 @@ package wegrus.clubwebsite.dto.member;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import wegrus.clubwebsite.entity.Member;
+import wegrus.clubwebsite.entity.member.Member;
 
 @Getter
 @AllArgsConstructor
