+++ src/main/java/wegrus/clubwebsite/entity/board/BoardCategory.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.board;
 
 public enum BoardCategory {
     GROUP, STUDY, BOARD
