+++ src/main/java/wegrus/clubwebsite/entity/board/BoardState.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.board;
 
 public enum BoardState {
     ACTIVATE, DEACTIVATE
