+++ src/main/java/wegrus/clubwebsite/entity/board/BoardType.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.board;
 
 public enum BoardType {
     IXPLOIT, IGDC, ALGROUS, WEBGROUS, PERSONAL, FREE, INFO, PROJECT, HOBBY, SUGGEST
