+++ src/main/java/wegrus/clubwebsite/entity/board/CommentLike.java
@@ -1,9 +1,10 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.board;
 
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import wegrus.clubwebsite.entity.member.Member;
 
 import javax.persistence.*;
