+++ src/main/java/wegrus/clubwebsite/entity/board/ReplyState.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.board;
 
 public enum ReplyState {
     ACTIVATE, DEACTIVATE
