+++ src/main/java/wegrus/clubwebsite/entity/member/MemberAcademicStatus.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.member;
 
 public enum MemberAcademicStatus {
     ATTENDING, ABSENCE, GRADUATED
