+++ src/main/java/wegrus/clubwebsite/entity/member/MemberGrade.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.member;
 
 public enum MemberGrade {
     FRESHMAN, SOPHOMORE, JUNIOR, SENIOR
