+++ src/main/java/wegrus/clubwebsite/entity/member/MemberRole.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity;
+package wegrus.clubwebsite.entity.member;
 
 public enum MemberRole {
     ROLE_GUEST, ROLE_CERTIFIED, ROLE_MEMBER, ROLE_MANAGER
