+++ src/main/java/wegrus/clubwebsite/repository/BoardRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.Board;
+import wegrus.clubwebsite.entity.board.Board;
 
 public interface BoardRepository extends JpaRepository<Board, Long> {
 }
