+++ src/main/java/wegrus/clubwebsite/repository/CommentLikeRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.CommentLike;
+import wegrus.clubwebsite.entity.board.CommentLike;
 
 public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
 }
