+++ src/main/java/wegrus/clubwebsite/repository/MemberRepository.java
@@ -2,7 +2,7 @@ package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import wegrus.clubwebsite.entity.Member;
+import wegrus.clubwebsite.entity.member.Member;
 
 import java.util.Optional;
