+++ src/main/java/wegrus/clubwebsite/repository/PostLikeRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.PostLike;
+import wegrus.clubwebsite.entity.board.PostLike;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
 }
