+++ src/main/java/wegrus/clubwebsite/repository/ReplyRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.Reply;
+import wegrus.clubwebsite.entity.board.Reply;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
 }
