+++ src/main/java/wegrus/clubwebsite/repository/ViewRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.View;
+import wegrus.clubwebsite.entity.board.View;
 
 public interface ViewRepository extends JpaRepository<View, Long> {
 }
