+++ src/main/java/wegrus/clubwebsite/service/MemberService.java
@@ -13,8 +13,8 @@ import wegrus.clubwebsite.dto.VerificationResponse;
 import wegrus.clubwebsite.dto.error.ErrorResponse;
 import wegrus.clubwebsite.dto.member.MemberAndJwtDto;
 import wegrus.clubwebsite.dto.member.MemberSignupRequest;
-import wegrus.clubwebsite.entity.Member;
-import wegrus.clubwebsite.entity.MemberRole;
+import wegrus.clubwebsite.entity.member.Member;
+import wegrus.clubwebsite.entity.member.MemberRole;
 import wegrus.clubwebsite.exception.MemberAlreadyExistException;
 import wegrus.clubwebsite.exception.MemberNotFoundException;
 import wegrus.clubwebsite.repository.MemberRepository;
